Add an input reading system
Create a unified interface for readers, with two fundamental methods,
peek(at) and consume(amt). Implement two readers. One which buffers at
most 2(may be subject to change) characters in memory, and one which
reads the contents of the input entirely in memory.

Create a binary crate for a soon-to-be command-line frontend and make
sure the two readers behave the same.