Make peeking readers idempotent as well as make consuming fallible
Change the reader interface to make peeking idempotent. Update both
current implementations. The changes to the in-memory reader are trivial,
but the changes to the buffered reader are more involved as they involve
a new field for storing the last character.

Change the reader interface to make consuming fallible. Again, update
both current implementations. The changes to the in-memory reader are
trivial, again. This interface change is because of the new needs of the
consume implementation for the buffered reader. The new implementation
calls the fill_buf() method, which is a fallible operation as reading
from the inner Read instance, or converting a slice of bytes to a string
slice might fail, those failures needing propagation. With this occasion
fill_buf()'s signature has been updated to use the appropiate Result
type.