Add a lexer
Implement a basic lexer that supports two kinds of tokens, with an
interface based on the one of the ReadInput trait(i.e., peek() and
consume()).
Implement Iterator for the lexer. It is left to be decided whether or
not this is preferable over a specialised method to create an iterator
over references to tokens, as the current implementation relies on peek
returning owned values, which is done using std::ptr::read. This is
abusable as peeking twice will return two owned values that are actually
the same value in memory.
Leave TODO notes