Make peeking a lexer safe and reimplement iterating for lexers
Return a reference to the current token instead of reading it, unsafely.
Make an intermediary for turning a lexer into an iterator. This ensures
that the iterator can't be consumed while iterating, since turning it
into an iterator takes ownership, as well as allowing for iterating over
owned tokens instead of references.